fix almost comments other than on utility type (Omit type)
