+++ 1_webbee_task_manager/src/components/TodoItem.tsx
@@ -1,5 +1,6 @@
-import React from 'react';
+import React, { useMemo, PureComponent, useState } from 'react';
 import MyButton from '../components/UI/button/MyButton';
+import { Todo } from '../Types';
 
 interface TodoItemProps {
   todo: Todo;
@@ -8,25 +9,32 @@ interface TodoItemProps {
   selectTodoIdForEdit: (id: Todo['id']) => void;
 }
 
-export const TodoItem: React.FC<TodoItemProps> = ({ todo, deleteTodo, checkTodo, selectTodoIdForEdit }) => (
-  <>
-    <div className="TodoItem">
+const TodoItem: React.FC<TodoItemProps> = ({ todo, deleteTodo, checkTodo, selectTodoIdForEdit }) => {
+  console.log('Render item ' + todo.id);
+
+  return (
+    <>
       <div className="TodoItem">
-        <div
-          className="TodoItem_text"
-          aria-hidden
-          style={{
-            textDecoration: todo.completed ? 'line-through' : 'none',
-          }}
-        >
-          <input type="checkbox" checked={todo.completed} onClick={() => checkTodo(todo.id)} />
-          {todo.title}
+        <div className="TodoItem">
+          {todo.id}
+          <div
+            className="TodoItem_text"
+            aria-hidden
+            style={{
+              textDecoration: todo.completed ? 'line-through' : 'none',
+            }}
+          >
+            <input type="checkbox" checked={todo.completed} onClick={() => checkTodo(todo.id)} />
+            {todo.title}
+          </div>
+        </div>
+        <div className="TodoItem">
+          <MyButton onClick={() => selectTodoIdForEdit(todo.id)}>Edit</MyButton>
+          <MyButton onClick={() => deleteTodo(todo.id)}>Delete</MyButton>
         </div>
       </div>
-      <div className="TodoItem">
-        <MyButton onClick={() => selectTodoIdForEdit(todo.id)}>Edit</MyButton>
-        <MyButton onClick={() => deleteTodo(todo.id)}>Delete</MyButton>
-      </div>
-    </div>
-  </>
-);
+    </>
+  );
+};
+
+export default React.memo(TodoItem);
