+++ 1_webbee_task_manager/src/components/TodoList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-
-import { TodoItem } from '../components/TodoItem';
-import { TodoPanel } from '../components/TodoPanel';
+import TodoItem from '../components/TodoItem';
+import TodoEditPanel from './TodoEditPanel';
+import { Todo } from '../Types';
 
 interface TodoListProps {
   todoIdForEdit: Todo['id'] | null;
@@ -9,11 +9,11 @@ interface TodoListProps {
   deleteTodo: (id: Todo['id']) => void;
   checkTodo: (id: Todo['id']) => void;
   selectTodoIdForEdit: (id: Todo['id']) => void;
-  changeTodo: ({ title }: Omit<Todo, 'id' | 'checked'>) => void;
+  changeTodo: (title: Omit<Todo, 'id' | 'checked'>) => void;
   cancelTodo: ({}: Omit<Todo, 'id' | 'checked'>) => void;
 }
 
-export const TodoList: React.FC<TodoListProps> = ({
+const TodoList: React.FC<TodoListProps> = ({
   todos,
   todoIdForEdit,
   changeTodo,
@@ -25,7 +25,9 @@ export const TodoList: React.FC<TodoListProps> = ({
   <div>
     {todos.map((todo) => {
       if (todo.id === todoIdForEdit)
-        return <TodoPanel mode="edit" changeTodo={changeTodo} editTodo={todo} cancelTodo={cancelTodo} />;
+        return (
+          <TodoEditPanel key={todo.id} mode="edit" changeTodo={changeTodo} editTodo={todo} cancelTodo={cancelTodo} />
+        );
       return (
         <TodoItem
           key={todo.id}
@@ -38,3 +40,5 @@ export const TodoList: React.FC<TodoListProps> = ({
     })}
   </div>
 );
+
+export default React.memo(TodoList);
