+++ 1_webbee_task_manager/src/css/app.css
@@ -21,6 +21,7 @@
 .TodoItem {
   display: flex;
   justify-content: space-between;
+  align-items: center;
 }
 
 .TodoItem_text {
